+++ python/paddle/fluid/tests/unittests/npu/test_conv3d_op_npu.py
@@ -364,7 +364,7 @@ class TestConv3DAPI(unittest.TestCase):
             dtype="float32",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NDHWC,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -375,7 +375,7 @@ class TestConv3DAPI(unittest.TestCase):
             data_format="NCDHW",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NCDHW,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -386,7 +386,7 @@ class TestConv3DAPI(unittest.TestCase):
             data_format="NCDHW",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NCDHW,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -397,7 +397,7 @@ class TestConv3DAPI(unittest.TestCase):
             data_format="NCDHW",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NDHWC,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -408,7 +408,7 @@ class TestConv3DAPI(unittest.TestCase):
             data_format="NDHWC",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NCDHW,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -419,7 +419,7 @@ class TestConv3DAPI(unittest.TestCase):
             data_format="NCDHW",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NCDHW,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -442,7 +442,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
 
         # ValueError: cudnn
         def run_1():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=3,
@@ -458,7 +458,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
 
         # ValueError: data_format
         def run_2():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=[3, 3, 3],
@@ -474,7 +474,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
 
         # ValueError: padding
         def run_3():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=3,
@@ -489,7 +489,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
         self.assertRaises(ValueError, run_3)
 
         def run_4():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=3,
@@ -504,7 +504,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
         self.assertRaises(ValueError, run_4)
 
         def run_5():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=0,
@@ -527,7 +527,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
         )
 
         def run_6():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=x,
                 num_filters=3,
                 filter_size=3,
@@ -543,7 +543,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
 
         # ValueError: groups
         def run_7():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=3,
@@ -559,7 +559,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
 
         # ValueError: filter num
         def run_8():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=0,
                 filter_size=0,
