+++ python/paddle/fluid/tests/unittests/test_conv3d_op.py
@@ -878,7 +878,7 @@ class TestConv3DAPI(unittest.TestCase):
             dtype="float32",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NDHWC,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -889,7 +889,7 @@ class TestConv3DAPI(unittest.TestCase):
             data_format="NCDHW",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NCDHW,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -900,7 +900,7 @@ class TestConv3DAPI(unittest.TestCase):
             data_format="NCDHW",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NCDHW,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -911,7 +911,7 @@ class TestConv3DAPI(unittest.TestCase):
             data_format="NCDHW",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NDHWC,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -922,7 +922,7 @@ class TestConv3DAPI(unittest.TestCase):
             data_format="NDHWC",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NCDHW,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -933,7 +933,7 @@ class TestConv3DAPI(unittest.TestCase):
             data_format="NCDHW",
         )
 
-        fluid.layers.conv3d(
+        paddle.static.nn.conv3d(
             input=input_NCDHW,
             num_filters=3,
             filter_size=[3, 3, 3],
@@ -956,7 +956,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
 
         # ValueError: cudnn
         def run_1():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=3,
@@ -972,7 +972,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
 
         # ValueError: data_format
         def run_2():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=[3, 3, 3],
@@ -988,7 +988,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
 
         # ValueError: padding
         def run_3():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=3,
@@ -1003,7 +1003,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
         self.assertRaises(ValueError, run_3)
 
         def run_4():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=3,
@@ -1018,7 +1018,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
         self.assertRaises(ValueError, run_4)
 
         def run_5():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=0,
@@ -1041,7 +1041,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
         )
 
         def run_6():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=x,
                 num_filters=3,
                 filter_size=3,
@@ -1057,7 +1057,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
 
         # ValueError: groups
         def run_7():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=3,
                 filter_size=3,
@@ -1073,7 +1073,7 @@ class TestConv3DAPI_Error(unittest.TestCase):
 
         # ValueError: filter num
         def run_8():
-            fluid.layers.conv3d(
+            paddle.static.nn.conv3d(
                 input=input,
                 num_filters=0,
                 filter_size=0,
