+++ python/paddle/fluid/tests/unittests/test_imperative_load_static_param.py
@@ -51,10 +51,10 @@ class TestDygraphLoadStatic(unittest.TestCase):
         conv3d_in = fluid.data(
             name='conv3d_in', shape=[None, 3, 12, 32, 32], dtype='float32'
         )
-        conv3d_out_1 = fluid.layers.conv3d(
+        conv3d_out_1 = paddle.static.nn.conv3d(
             input=conv3d_in, num_filters=2, filter_size=3, act="relu"
         )
-        conv3d_out_2 = fluid.layers.conv3d(
+        conv3d_out_2 = paddle.static.nn.conv3d(
             input=conv3d_in, num_filters=2, filter_size=3, act="relu"
         )
