+++ python/paddle/static/nn/__init__.py
@@ -14,6 +14,7 @@
 
 from .common import fc  # noqa: F401
 from .common import deform_conv2d  # noqa: F401
+from .common import conv3d  # noqa: F401
 
 from ...fluid.layers import batch_norm  # noqa: F401
 from ...fluid.layers import bilinear_tensor_product  # noqa: F401
@@ -21,7 +22,6 @@ from ...fluid.layers import case  # noqa: F401
 from ...fluid.layers import cond  # noqa: F401
 from ...fluid.layers import conv2d  # noqa: F401
 from ...fluid.layers import conv2d_transpose  # noqa: F401
-from ...fluid.layers import conv3d  # noqa: F401
 from ...fluid.layers import conv3d_transpose  # noqa: F401
 from ...fluid.layers import create_parameter  # noqa: F401
 from ...fluid.layers import crf_decoding  # noqa: F401
